# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: $MAIL — Solana wallet mailbox API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a folder for the signed-in wallet
      description: Listing mailbox/requests/junk marks newly-surfaced incoming mail as delivered.
      parameters:
        - name: folder
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Folder"
      responses:
        "200":
          description: Messages in the folder, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send mail to any valid Solana wallet address
      description: The recipient does not need to be a $MAIL user. The message is recorded as addressed to that wallet.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Mail addressed to the recipient wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Invalid recipient address or empty body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "403":
          description: Recipient has blocked this sender (RETURN TO SENDER)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /messages/{id}:
    get:
      operationId: getMessage
      tags: [messages]
      summary: Get a single message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /messages/{id}/thread:
    get:
      operationId: getMessageThread
      tags: [messages]
      summary: Get the full correspondence thread containing this message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Thread messages, oldest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /messages/{id}/open:
    post:
      operationId: openMessage
      tags: [messages]
      summary: Mark a message opened by the authenticated recipient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The opened message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /messages/{id}/accept:
    post:
      operationId: acceptMessage
      tags: [messages]
      summary: Accept mail into the Mailbox (from Requests or Junk)
      description: Also establishes the sender as approved, so future mail from them lands in the Mailbox.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The accepted message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /messages/{id}/junk:
    post:
      operationId: junkMessage
      tags: [messages]
      summary: Move a message to Junk Mail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The junked message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /messages/{id}/share:
    get:
      operationId: getShareCard
      tags: [messages]
      summary: Get privacy-first share card data for a delivery
      description: Message content stays PRIVATE unless the owner explicitly requests reveal=true.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Share card data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareCard"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    Folder:
      type: string
      enum: [mailbox, requests, junk, sent]
    MessageStatus:
      type: string
      description: "addressed: recorded for the recipient wallet. delivered: surfaced on a $MAIL surface the recipient accessed. opened: the authenticated recipient opened it."
      enum: [addressed, delivered, opened]
...
    Message:
      type: object
      properties:
        id:
          type: string
        senderWallet:
          type: string
        recipientWallet:
          type: string
        body:
          type: string
        folder:
          $ref: "#/components/schemas/Folder"
        status:
          $ref: "#/components/schemas/MessageStatus"
        senderVerified:
          type: boolean
          description: True when the sender proved wallet ownership by signature at send time
        containsUrl:
          type: boolean
        urlDomains:
          type: array
          items:
            type: string
          description: Destination domains detected in the body, for link-safety warnings
        threadId:
          type: string
        replyToMessageId:
          type: ["string", "null"]
        createdAt:
          type: string
        deliveredAt:
          type: ["string", "null"]
        openedAt:
          type: ["string", "null"]
        isOutgoing:
          type: boolean
          description: True when the signed-in wallet is the sender
      required:
        - id
        - senderWallet
        - recipientWallet
        - body
        - folder
        - status
        - senderVerified
        - containsUrl
        - urlDomains
        - threadId
        - replyToMessageId
        - createdAt
        - deliveredAt
        - openedAt
        - isOutgoing
    MessageInput:
      type: object
      properties:
        recipientWallet:
          type: string
          description: Any valid base58 Solana wallet address — recipient does not need to be a $MAIL user
        body:
          type: string
          minLength: 1
          maxLength: 4000
        replyToMessageId:
          type: string
          description: Set when replying to keep correspondence in one thread
      required:
        - recipientWallet
        - body
    ShareCard:
      type: object
      properties:
        messageId:
          type: string
        headline:
          type: string
          description: e.g. "I GOT MAIL"
        deliveryType:
          type: string
          description: e.g. "STANDARD MAIL", "SIGNED MAIL"
        senderVerified:
          type: boolean
        receivedAt:
          type: string
        revealed:
          type: boolean
          description: False by default. The message stays PRIVATE unless the owner explicitly reveals it.
        body:
          type: ["string", "null"]
          description: Message content — null unless explicitly revealed by the owner
      required:
        - messageId
        - headline
        - deliveryType
        - senderVerified
        - receivedAt
        - revealed
        - body
...