# lib/api-zod/src/generated/types/folder.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * $MAIL — Solana wallet mailbox API
 * OpenAPI spec version: 0.1.0
 */

export type Folder = typeof Folder[keyof typeof Folder];


export const Folder = {
  mailbox: 'mailbox',
  requests: 'requests',
  junk: 'junk',
  sent: 'sent',
} as const;