# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * $MAIL — Solana wallet mailbox API
 * OpenAPI spec version: 0.1.0
 */

export interface MessageInput {
  /** Any valid base58 Solana wallet address — recipient does not need to be a $MAIL user */
  recipientWallet: string;
  /**
     * @minLength 1
     * @maxLength 4000
     */
  body: string;
  /** Set when replying to keep correspondence in one thread */
  replyToMessageId?: string;
}