# lib/api-zod/src/generated/types/shareCard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * $MAIL — Solana wallet mailbox API
 * OpenAPI spec version: 0.1.0
 */

export interface ShareCard {
  messageId: string;
  /** e.g. "I GOT MAIL" */
  headline: string;
  /** e.g. "STANDARD MAIL", "SIGNED MAIL" */
  deliveryType: string;
  senderVerified: boolean;
  receivedAt: string;
  /** False by default. The message stays PRIVATE unless the owner explicitly reveals it. */
  revealed: boolean;
  /**
     * Message content — null unless explicitly revealed by the owner
     * @nullable
     */
  body: string | null;
}